requirements.txt
anthropic
pydantic
yfinance
python-dotenv
mcp
langgraph
langchain-core
langchain_anthropic
fastapi
sqlalchemy
psycopg2-binary
uvicorn
python-multipart
